% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example_researcher_dataset.R
\name{example_researcher}
\alias{example_researcher}
\title{Example data set from publication for scatter plot and Bland Altman plot}
\description{
Contains the data set (\code{example_researcher}).
}
\keyword{data}

