% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_indicators.R
\name{point_indicators}
\alias{point_indicators}
\title{point_indicators}
\usage{
point_indicators(indicators, size = 1, vjust = 0.5)
}
\arguments{
\item{indicators}{data.frame whith thre variables: 1 actors, 2 indicators and 3 values}

\item{size}{size of value labels}

\item{vjust}{vertical justification of value labels}
}
\description{
point_indicators
}
